WIP: Disable GIL: Implement free threading for Python 3.13
This works now completely with 3.13t. There are no more segfaults.

The speed factor is slightly better than 4 with a 10 core
Max Studio M1. This was much better with nogil. It will improve
in the next versions.

Currently, the acceleration works on Python 3.13.0t but has a bug
on 3.13.1t and 3.13.2t.

The basic functionality is there, thread-safety is a matter for
the follow-up patch.

Task-number: PYSIDE-2221
Change-Id: Ia560b6a27a8cd931c7a0f9bc0e348c10fde270a1